tests/wpt/web-platform-tests/websockets/interfaces/WebSocket/readyState/005.html

<!doctype html>
<title>WebSockets: defineProperty setter for readyState</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js?pipe=sub></script>
<meta name="variant" content="">
<meta name="variant" content="?wss">
<div id=log></div>
<script>
test(function(){
  window.setter_ran = false;
  Object.defineProperty(WebSocket.prototype, 'readyState', {
    set: function(v) { window[v] = true; }
  });
  var ws = new WebSocket('ws://example.invalid/');
  ws.readyState = 'setter_ran';
  assert_true(setter_ran);
});
</script>
